So we're going to do this really object oriented.  Doing his, we'll be able to avoid circular logic.
So it's going to be expressions.  Every expression has an array of cas objects.  Exponents are going to contain an expression, and be separate from the cas object.

So expressoins contain exponents, exponents contain expressions? Sounds like some circular voodoo, right? Wrong.  Were going to make an abstract expression type.  Real expressions are going to contain an exponent, but were going to have endexpressoins that don't.

Having this dichotomy should make it a bit easier to avoid if statements, and relatively esy to parse